Rename FioBankService to BankTransactionService for abstraction
