+++ bounded-contexts/budget/src/main/scala/works/iterative/incubator/budget/domain/service/BankTransactionService.scala
@@ -4,13 +4,16 @@ import works.iterative.incubator.budget.domain.model.{AccountId, Transaction}
 import java.time.LocalDate
 import zio.*
 
-/** Service interface for importing transactions from Fio Bank API.
+/** Service interface for importing transactions from bank APIs.
+  *
+  * This service defines the contract for fetching transactions from any bank's API.
+  * Specific bank implementations will provide concrete implementations of this interface.
   *
   * Category: Service Interface
   * Layer: Domain
   */
-trait FioBankService:
-  /** Fetches transactions from Fio Bank API for a specified date range.
+trait BankTransactionService:
+  /** Fetches transactions from a bank API for a specified date range.
     *
     * @param accountId
     *   The ID of the source account to fetch transactions for
@@ -29,7 +32,7 @@ trait FioBankService:
 
 /** Companion object providing ZIO accessor methods.
   */
-object FioBankService:
+object BankTransactionService:
   /** Accesses the service to fetch transactions.
     *
     * @param accountId
@@ -39,12 +42,12 @@ object FioBankService:
     * @param endDate
     *   The end date of the range
     * @return
-    *   A ZIO effect that requires FioBankService and returns a list of transactions or a
+    *   A ZIO effect that requires BankTransactionService and returns a list of transactions or a
     *   service-specific error
     */
   def fetchTransactions(
       accountId: AccountId,
       startDate: LocalDate,
       endDate: LocalDate
-  ): ZIO[FioBankService, Throwable, List[Transaction]] =
+  ): ZIO[BankTransactionService, Throwable, List[Transaction]] =
     ZIO.serviceWithZIO(_.fetchTransactions(accountId, startDate, endDate))
