+++ bounded-contexts/budget/src/test/scala/works/iterative/incubator/budget/ui/transaction_import/TransactionImportPresenterLiveSpec.scala
@@ -2,7 +2,6 @@ package works.iterative.incubator.budget.ui.transaction_import
 
 import works.iterative.incubator.budget.domain.model.AccountId
 import works.iterative.incubator.budget.domain.service.*
-import works.iterative.incubator.budget.domain.service.FioBankServiceNormal
 import works.iterative.incubator.budget.infrastructure.adapter.*
 import works.iterative.incubator.budget.ui.transaction_import.models.*
 import java.time.LocalDate
@@ -88,7 +87,7 @@ object TransactionImportPresenterLiveSpec extends ZIOSpecDefault:
         result <- TransactionImportPresenter.validateAndProcess(command)
       yield assert(result.isRight)(isTrue) &&
         assert(result.toOption.map(_.transactionCount).getOrElse(0))(
-          equalTo(TestFioBankService.TestTransactionCount)
+          equalTo(TestBankTransactionService.TestTransactionCount)
         )
     },
